Create Bayesian_hierarchical
These codes are from Lützen Portengen and Calvin Ge on a hierarchical Bayesian model for agnostic analysis of occupation-related lung cancer risks in the SYNERGY case-control population. The first model (M1) is a is a simple glm with all 5-digit ISCO-68 jobs and all covariates. The estimates and variances from M1 are in turn used in M2, which is a Bayesian model estimating disease risks at the 5-digit and 3-digit ISCO-68 levels.